+++ sinch/domains/voice/api/v2/sinch_events.py
@@ -9,7 +9,7 @@ from sinch.core.internal.sinch_events.utils import (
 )
 from sinch.core.models.internal.utils import strip_unset
 from sinch.core.sentinel import UNSET, Unset, UnsetOr
-from sinch.domains.voice.models.v2.sinch_events.incoming_call_events import (
+from sinch.domains.voice.models.v2.svaml.shared.incoming_call_response_events import (
     IncomingCallResponseEvents,
 )
 from sinch.domains.voice.models.v2.sinch_events.voice_sinch_event_request import (