+++ sinch/domains/voice/models/v2/sinch_events/__init__.py
@@ -1,9 +1,6 @@
 from sinch.domains.voice.models.v2.sinch_events.event_type import (
     EventType,
 )
-from sinch.domains.voice.models.v2.sinch_events.incoming_call_events import (
-    IncomingCallResponseEvents,
-)
 from sinch.domains.voice.models.v2.sinch_events.menu_input import MenuInput
 from sinch.domains.voice.models.v2.sinch_events.voice_sinch_event_request import (
     VoiceSinchEventRequest,
@@ -17,5 +14,4 @@ __all__ = [
     "EventType",
     "VoiceSinchEventRequest",
     "VoiceSinchEventResponse",
-    "IncomingCallResponseEvents",
 ]