+++ sinch/domains/voice/models/v2/svaml/types/incoming_call_response_events_dict.py
@@ -1,0 +1,12 @@
+from typing import TYPE_CHECKING, List, TypedDict
+
+from typing_extensions import NotRequired
+
+if TYPE_CHECKING:
+    from sinch.domains.voice.models.v2.svaml.types.svaml_command_dict import (
+        SvamlCommandDict,
+    )
+
+
+class IncomingCallResponseEventsDict(TypedDict):
+    on_hangup: NotRequired[List["SvamlCommandDict"]]