+++ sinch/domains/voice/models/v2/svaml/types/svaml_input_dict.py
@@ -1,7 +1,10 @@
 from typing import List
 
-from typing_extensions import TypedDict
+from typing_extensions import NotRequired, TypedDict
 
+from sinch.domains.voice.models.v2.svaml.types.incoming_call_response_events_dict import (
+    IncomingCallResponseEventsDict,
+)
 from sinch.domains.voice.models.v2.svaml.types.svaml_command_dict import (
     SvamlCommandDict,
 )
@@ -9,3 +12,5 @@ from sinch.domains.voice.models.v2.svaml.types.svaml_command_dict import (
 
 class SvamlInputDict(TypedDict):
     commands: List[SvamlCommandDict]
+    call_name: NotRequired[str]
+    events: NotRequired[IncomingCallResponseEventsDict]